chore: Refactor hook execution. (#417)
This code lifts the hook running code out of the client.
